requirements.txt
alembic==0.9.6
bcrypt==3.1.4
click==6.7
coverage==4.4.2
eventlet==0.21.0
flask>=0.12.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2>=2.10.1
jsonschema==2.6.0
Mako==1.0.7
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
selenium==3.8.1
six==1.11.0
SQLAlchemy>=1.3.0
Werkzeug==0.15.5
Flask-Cognito
boto3
pytest-flask